third_party/WebKit/LayoutTests/crypto/clone-rsaKey-public-expected.txt

Tests structured cloning of RSA public keys (without a hash)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is ""
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: ff063f004b03040180080301000101a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a41020301000100
PASS: Cloned key exported data should be [30819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a410203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is ""
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: ff063f004b03040180100301000101a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef020301000100
PASS: Cloned key exported data should be [30820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef0203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: ff063f004b03040180080301000103a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a41020301000100
PASS: Cloned key exported data should be [30819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a410203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: ff063f004b03040180100301000103a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef020301000100
PASS: Cloned key exported data should be [30820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef0203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: ff063f004b03040180080301000145a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a41020301000100
PASS: Cloned key exported data should be [30819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a410203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: ff063f004b03040180100301000145a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef020301000100
PASS: Cloned key exported data should be [30820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef0203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: ff063f004b030401800803010001c301a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a410203010001
PASS: Cloned key exported data should be [30819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a410203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: ff063f004b030401801003010001c301a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef0203010001
PASS: Cloned key exported data should be [30820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef0203010001] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is ""
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: ff063f004b03040180080301000100a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a41020301000100

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is ""
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: ff063f004b03040180100301000100a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef020301000100

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: ff063f004b03040180080301000102a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a41020301000100

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: ff063f004b03040180100301000102a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef020301000100

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: ff063f004b03040180080301000144a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a41020301000100

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: ff063f004b03040180100301000144a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef020301000100

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 1024
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 1024
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: ff063f004b030401800803010001c201a20130819f300d06092a864886f70d010101050003818d0030818902818100b289c62ecc3ddf64154817203439eaa0dc07a65954429a7b6098a77235673d96df1f06bd3c1ae73990867199e678bf95b3728fcd4686136e6ee9dd4c09eb490eb7cb953c388668b759263f61d6a7dfcabf27b5c9d6972455b12b66d483843286d6b871f35f912a773c97c1932255fcee05ce7b8af213879f017de4232a306a410203010001

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS importedKey.algorithm.modulusLength is 2048
PASS: importedKey.algorithm.publicExponent should be [010001] and was
PASS importedKey.algorithm.hash is undefined.
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey != clonedKey is true
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "RSAES-PKCS1-v1_5"
PASS clonedKey.algorithm.modulusLength is 2048
PASS: clonedKey.algorithm.publicExponent should be [010001] and was
PASS clonedKey.algorithm.hash is undefined.
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: ff063f004b030401801003010001c201a60230820122300d06092a864886f70d01010105000382010f003082010a0282010100b4c8b631194aef0154b1479ab7a534b60ca878981108680f0ae6b7c88cb6010f6dbf9f665646208410575cb923b26bf874a24b4cd801c9bba967062ae506cdcf307add4380d0d93077a4c1f0fc06d498dc729f811335c530b90fe9bf9f3979ccec050a48e8923045b19368e1e89ea4157538e8059e53320f47c155f1741310a93ed153a7f3af2d46c388b95d82518527a02b7bd9ab7edc4bcb737677f679c5c6de5e1ebed3a29d6b99b8eace2d59ceb533e001cf39c5671495d51d3ee34406ea4fdb0c626dee68da256b8a12f9f65059ccc85a2190ce1385152d62785e00cae702e77c4c597b86a6268adbf043dda68881c790f1517671fb7d198fca5ba97bef0203010001

PASS successfullyParsed is true

TEST COMPLETE
